fix: resolve Clippy warnings
- Remove unnecessary i64 cast in search.rs
- Use strip_prefix() instead of manual slicing in summarizers.rs
- Replace map_or with is_some_and() for cleaner code
- Use Ok() pattern instead of .ok() in templates.rs
- All Clippy warnings resolved (--workspace -- -D warnings passes)

+++ src-tauri/src/commands/search.rs
@@ -84,7 +84,7 @@ pub async fn get_embedding_status(
         .repo()
         .count_notes()
         .await
-        .map_err(|e| CommandError::Vault(e.to_string()))? as i64;
+        .map_err(|e| CommandError::Vault(e.to_string()))?;
 
     Ok(EmbeddingStatus {
         connected,
